requirements.txt
Pillow>=2.8.0
